requirements.txt
snowflake-snowpark-python
pandas
numpy
streamlit
matplotlib
altair
prophet
statsmodels
holidays
numpy
pandas-datareader
cython
future
decorator
scikit-learn
tabulate
pillow